+++ ios/Tests/ByteowlsCapacitorFilesharerTests/ByteowlsCapacitorFilesharerTests.swift
@@ -1,8 +1,0 @@
-import XCTest
-@testable import ByteowlsCapacitorFilesharer
-
-final class ByteowlsCapacitorFilesharerTests: XCTestCase {
-    func testPluginCanBeCreated() {
-        XCTAssertNotNil(FileSharerPlugin())
-    }
-}